# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Direct Host-to-Host Bank Payment Initiation, Webhook Ingestion, and Atomic Ledger Reconciliation Engine for DOKI.",
        "title": "DOKI Direct Banking \u0026 Billing Engine API",
        "contact": {
            "name": "DOKI Fintech Engineering Team",
            "email": "fintech@doki.travel"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/admin/properties/{property_id}/payments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[READ] Paginated ledger listing for hotel accountants, filterable by status, bank_id, payment_method, PNR, and date range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "List and filter property payment ledger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property UUID",
                        "name": "property_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Payment Status filter (INITIATED, SETTLED, VOIDED, REFUNDED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bank ID filter (e.g. BANK_CBE, BANK_BOA)",
                        "name": "bank_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Payment Method filter (DIRECT_BANK, CASH, POS)",
                        "name": "payment_method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Booking PNR filter",
                        "name": "pnr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Reference Number filter",
                        "name": "reference_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size limit (default: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated ledger records",
                        "schema": {
                            "$ref": "#/definitions/billing.PaginatedPaymentsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid property UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/properties/{property_id}/payments/manual": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[CREATE] Records manual counter cash or terminal POS settlement directly into the guest folio and confirms reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Record front-desk cash or POS payment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property UUID",
                        "name": "property_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Manual payment settlement payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.ManualPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Manual payment recorded and reservation confirmed",
                        "schema": {
                            "$ref": "#/definitions/billing.ManualPaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payments/initiate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[CREATE] Generates reference_number, 6-character PNR, stay duration, and bank payment instructions for a held reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Initiate direct bank payment order",
                "parameters": [
                    {
                        "description": "Payment initiation payload containing reservation_id",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/billing.InitiatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment order created successfully",
                        "schema": {
                            "$ref": "#/definitions/billing.InitiatePaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload or malformed UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid authorization token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Reservation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Reservation not in HELD state or order already settled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/payments/{payment_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[READ] Fetch complete payment order details, PNR, payer account, bank reference, and ledger transaction history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment order profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment Order UUID",
                        "name": "payment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment order details",
                        "schema": {
                            "$ref": "#/definitions/billing.PaymentOrderDetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payment UUID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Payment order not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "billing.InitiatePaymentRequest": {
            "type": "object",
            "required": [
                "reservation_id"
            ],
            "properties": {
                "reservation_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "billing.ManualPaymentRequest": {
            "type": "object",
            "required": [
                "amount_minor",
                "payment_method",
                "receipt_number",
                "reservation_id"
            ],
            "properties": {
                "amount_minor": {
                    "type": "integer",
                    "example": 450000
                },
                "notes": {
                    "type": "string",
                    "example": "Front desk cash counter settlement"
                },
                "payment_method": {
                    "type": "string",
                    "enum": [
                        "CASH",
                        "POS",
                        "CARD",
                        "MANUAL"
                    ],
                    "example": "CASH"
                },
                "receipt_number": {
                    "type": "string",
                    "example": "REC-2026-001"
                },
                "reservation_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "staff_id": {
                    "type": "string",
                    "example": "770e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "billing.PaginatedPaymentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/billing.PaymentOrderDetailsResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 20
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 125
                },
                "total_pages": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}